Move `remote` one level up
