+++ src/query/remote/static_resolver.go
