+++ src/query/storage/remote/storage.go
@@ -26,8 +26,8 @@ import (
 
 	"github.com/m3db/m3/src/query/block"
 	"github.com/m3db/m3/src/query/errors"
+	"github.com/m3db/m3/src/query/remote"
 	"github.com/m3db/m3/src/query/storage"
-	"github.com/m3db/m3/src/query/tsdb/remote"
 )
 
 // Options contains options for remote clients.
